feat: redesign DownloadButton for audio and file attachment widgets
