+++ src/components/Attachment/components/index.ts
@@ -1,3 +1,2 @@
-export * from './AudioDownloadButton';
 export * from './DownloadButton';
 export * from './FileSizeIndicator';
