+++ src/components/Attachment/styling/index.scss
@@ -2,6 +2,7 @@
 @use 'AttachmentActions';
 @use 'Audio';
 @use 'CardAudio';
+@use 'DownloadButton';
 @use 'Geolocation';
 @use 'Giphy';
 @use 'LinkPreview';
